requirements.txt
langchain
langchain-core

langchain-google-genai
google-generativeai

python-dotenv